% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{Dmisc-package}
\alias{Dmisc-package}
\title{Dieter's misc functions, with focus on LaTeX table printout.}
\description{
Functions to export lme results and ODBC tables to LaTeX.
}
\details{
\tabular{ll}{ Package: \tab Dmisc\cr Type: \tab Package\cr Version: \tab
1.0\cr Date: \tab 2006-01-03\cr License: \tab GPL\cr }
}
\author{
Dieter Menne, \email{dieter.menne@menne-biomed.de}
}
\keyword{misc}
\keyword{print}

