% Generated by roxygen2 (4.0.1): do not edit by hand
\name{DMlatexSummaryby}
\alias{DMlatexSummaryby}
\title{Print LaTeX tables of summary statistics broken down by factors.}
\usage{
DMlatexSummaryby(object, title = "", INDICES = NULL, file = "",
  caption = "\%s", label = "", ctable = FALSE, rowlabel = "",
  moredec = 0, where = "!htbp", onetable = NULL, lang = "en",
  noQuantiles = FALSE, ...)
}
\arguments{
\item{object}{data frame with numerical and factor variables}

\item{title}{to be printed in upper left corner}

\item{INDICES}{a factor or list of factors for breakdown. Defaults to no
breakdown, resulting in one detail table only.}

\item{file}{output file name; default prints output to the standard output,
which is the choice for Sweave.}

\item{caption}{text for caption, with sprintf substitions.  Defaults to
factor combination, followed by explanation of the column headers.}

\item{label}{if no label given, label of table, defaults to
\code{tab:<factorname1><factor1>:<factornameN><factorN>.} If not "", and
there is no colon in the label, \code{tab:<label>.<factorname1>...} is used
as table label.  If there is a colon in the label, the label is used as is.}

\item{ctable}{Uses \code{ctable}-formatting of LaTeX by default.}

\item{rowlabel}{optional text for left table header cell.}

\item{moredec}{Show more decimals than by default. Fractional values, e.g.
0.5 can be used here such that boundary case are rounded up instead of down}

\item{where}{positioning parameter for LaTeX}

\item{onetable}{If TRUE, force multiple tables into one.  If FALSE, print
one table per combination of factors.  Internally set to TRUE when there is
only one numerical factor.}

\item{lang}{Language, currently \code{en} (default) or \code{de}.}

\item{noQuantiles}{Suppress quantiles if TRUE}

\item{\dots}{Additional parameters passed to \code{latex} in \code{Hmisc}.}
}
\value{
Returns the result of the call to \code{latex} in Hmisc.
}
\description{
LaTex output of summary statistics of a data frame. Prints one line for
minimum, maximum median, mean, quartiles, number of elements, standard
deviation and standard error for each numerical variable.  A separate table
is written for each unique combinations of factors. Labeling and generation
of captions is automatic.
}
\note{
Requires \code{ctable} in your Snw/tex file. In a
multi-table breakdown, only the caption of the first table has the
explanation of the headers.
}
\examples{
# one table only
DMlatexSummaryby(warpbreaks)
# a table for each combination
DMlatexSummaryby(warpbreaks,INDICES=c("wool","tension"))
DMlatexSummaryby(warpbreaks,INDICES=c("wool","tension"),noQuantiles=TRUE,
          caption="Warpbreaks \%s At the end", file="")
DMlatexSummaryby(warpbreaks,INDICES=c("wool","tension"),
          caption="Warpbreaks \%s At the end", file="",onetable=TRUE)
}
\author{
Dieter Menne, \email{dieter.menne@menne-biomed.de}
}
\seealso{
\code{\link[Hmisc]{latex}}, \code{\link{by}}
}
\keyword{print}

