% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/latexODBCTable.R
\name{latexODBCTable}
\alias{latexODBCTable}
\title{Retrieve database table and format it nicely for LaTeX}
\usage{
latexODBCTable(object, title = "", table = NULL, caption = NULL,
  longtable = FALSE, ctable = FALSE, label = NULL, file = "",
  sql = paste("select * from", table), where = "!htbp", mincapwidth = 120,
  width, as.is = FALSE, ...)
}
\arguments{
\item{object}{path of Excel or Access file. Currently, no other ODBC source
are supported, but changing this should be easy.}

\item{title}{to be printed in upper left corner}

\item{table}{name of table or query in database, or named range in Excel.
In theory, giving worksheet names for Excel should work, but results are
mixed, so better always use named ranges. Not used if sql is given.}

\item{caption}{optional caption. Default is \emph{Table <tablename>,
n=<number of rows>}}

\item{longtable}{use longtable; cannot be combinded with \code{ctable}}

\item{ctable}{Uses \code{ctable}-formatting of LaTeX by default.}

\item{label}{optional label. Default is \emph{tab:tablename}}

\item{file}{output file name; default prints output to the standard output,
which is the choice for \emph{Sweave}.}

\item{sql}{SQL query to retrieve data. Default is whole table.  Parameter
\code{table} is not used if query is given.}

\item{where}{positioning parameter for LaTeX}

\item{mincapwidth}{for ctable, the minimum caption width in mm. Requires
ctable1.sty or a more recent update of ctable.}

\item{width}{for ctable, width of table. Should be given when column type
\code{X} is used.}

\item{as.is}{passed to sqlQuery; if true, string are not converted to
factors.}

\item{\dots}{Additional parameters passed to \code{latex} in \code{Hmisc}.}
}
\value{
Returns the table retrieved from the database. This is somewhat
inconsistent to \code{latex.lme} and friends that return the latex object,
but it's convenient if single items of the table are to be displayed in
Sweave.
}
\description{
Generates LaTeX table from data retrieved from an Excel or Access file.
The table is closed immediately after retrieval.
}
\note{
Requires \code{ctable} in your Snw/tex file if ctable = TRUE
}
\examples{
\dontrun{
   # tex output in Sweave
   database = system.file("extdata", "testtable.accdb", package = "Dlatex")
   tb <- latexODBCTable(database, table = "Subject")


}
}
\author{
Dieter Menne, \email{dieter.menne@menne-biomed.de}
}
\seealso{
\code{\link{lme.value}}, \code{\link[Hmisc]{latex}},
\code{\link{latex.lme}}, \code{\link[nlme]{lme}}
}
\keyword{print}

