% Generated by roxygen2 (4.0.1): do not edit by hand
\name{removeFormFunc}
\alias{removeFormFunc}
\title{Lexically removes a function from a formula.}
\usage{
removeFormFunc(formula, func = "I")
}
\arguments{
\item{formula}{valid formula object}

\item{func}{Formula to be remove, without the \code{()}}
}
\value{
Returns the string without the formula part.
}
\description{
Remove the I(..) or similar function from a formula. So
\code{effect ~ I(sqrt(nitro))*treat + I(nitro^2)}
becomes \code{effect ~ sqrt(nitro)*treat + nitro^2}
}
\examples{
removeFormFunc(effect ~ I(sqrt(nitro))*treat + I(nitro^2))
}
\author{
Thomas Lumley, Dieter Menne
}
\references{
After Thomas Lumley
\url{http://finzi.psych.upenn.edu/R/Rhelp02a/archive/30590.html} and a tip
from Gabor Grothendiek.
}

