% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/removeFormFunc.R
\name{removeFormFunc}
\alias{removeFormFunc}
\title{Lexically removes a function from a formula.}
\usage{
removeFormFunc(formula, func = "I")
}
\arguments{
\item{formula}{valid formula object}

\item{func}{Formula to be remove, without the \code{()}}
}
\value{
Returns the string without the formula part.
}
\description{
Remove the I(..) or similar function from a formula. So
\code{effect ~ I(sqrt(nitro))*treat + I(nitro^2)}
becomes \code{effect ~ sqrt(nitro)*treat + nitro^2}
}
\examples{
removeFormFunc(effect ~ I(sqrt(nitro))*treat + I(nitro^2))
}
\author{
Thomas Lumley, Dieter Menne
}

